Rename legacy code to classic, since this will be the classic version of CavSim
