+++ src/cpp/CMakeLists.txt
@@ -4,22 +4,22 @@ add_library(${PROJECT_NAME}
             STATIC
             Cavity.cpp
             Cavity.hpp
-            Legacy/solver.cpp
-            Legacy/solver.hpp
-            Legacy/numeric.cpp
-            Legacy/numeric.hpp
-            Legacy/IO.cpp
-            Legacy/IO.hpp
-            Legacy/Structures.cpp
-            Legacy/Structures.hpp
-            Legacy/WUDS.cpp
-            Legacy/WUDS.hpp
-            Legacy/PRIMEcorrection.cpp
-            Legacy/PRIMEcorrection.hpp
-            Legacy/MassEquation.cpp
-            Legacy/MassEquation.hpp
-            Legacy/MomentumEquation.cpp
-            Legacy/MomentumEquation.hpp)
+            classic/solver.cpp
+            classic/solver.hpp
+            classic/numeric.cpp
+            classic/numeric.hpp
+            classic/IO.cpp
+            classic/IO.hpp
+            classic/Structures.cpp
+            classic/Structures.hpp
+            classic/WUDS.cpp
+            classic/WUDS.hpp
+            classic/PRIMEcorrection.cpp
+            classic/PRIMEcorrection.hpp
+            classic/MassEquation.cpp
+            classic/MassEquation.hpp
+            classic/MomentumEquation.cpp
+            classic/MomentumEquation.hpp)
 
 add_executable (${PROJECT_NAME}Old main.cpp)
