+++ src/cpp/classic/MassEquation.cpp
