+++ src/cpp/classic/MassEquation.hpp
