+++ src/cpp/classic/solver.cpp
