+++ src/cpp/classic/solver.hpp
